Schedule repo updates in background on package visit.
Resolves #29.

When visiting a package that is contained by a repo, schedule the repo
to be updated in background. Debounce to 20 seconds. Limit queue to 10.

This way, the repository will be updated for next visitor, or the same
visitor if they refresh.

A future enhancement is to display a yellow notification if an update
is detected after initial page load (see #30).