frontend/selectlistview: Remove Cmd+R, R shortcuts (keep only F).
The motivation for Cmd+R (and R, due to some browsers not supporting
Cmd+R overrides) was to use same shortcut as in Sublime Text, which I'm
used to.

However, ST is an editor, so it requires the use of a modifier such as
Cmd. Being a read-only code view, gotools.org can afford to skip the
modifier.

After godoc.org has chosen to to use F shortcut, I've added it to
gotools.org to stay consistent. Since then, I've started to use and
rely on F shortcut most of the time, and find it more natural.

Remove Cmd+R and R shortcuts now that we can do it for gotools.org
only. That way, there's only one canonical way of accessing the select
list view dropdown and jumping to an identifier, and it's the same as
on godoc.org.

This further increases the priority of #15.