Remove support for local GOPATH packages.
Previously, there were two primary use cases for gtdo binary. One was
running it in production mode at https://gotools.org. The other was to
run it locally on one's computer.

To support the latter, there was code to serve packages from local
GOPATH instead of always using vcs store.

This duality added complexity and made testing, development harder.
Right now, the code is in poor shape and needs to be improved. It
helps to reduce scope temporarily, and remove the local use case,
leaving only support for https://gotools.org.

Later on, this functionality can be brought back (perhaps via a
different binary), and done in a way that is in clean, easy to read,
maintain, and develop.

Doing this change now also fixes a bug, where visiting a non-existing
revision of a standard library package would give a misleading error.
E.g., visiting https://gotools.org/bytes?rev=xyz would previously give:

	unrecognized import path "bytes"

Now it says revision not found. When restoring GOPATH functionality in
the future, this bug should be easy to avoid.