Display only vertical scrollbar in TOC as needed.
The previous overflow: scroll; style made both the vertical and
horizontal scrollbars always visible when the OS setting is to always
show scrollbars (rather than overlay them only when scrolling).

Instead, we want to use something closer to:

	overflow-y: auto;
	overflow-x: hidden;

Which makes the vertical scrollbar appear only when needed, and the
horizontal scrollbar never appears. When overflow-x: hidden; is set,
overflow-y: visible;, which is the default, behaves the same as
overflow-y: auto;, so the CSS can be simplified to just:

	overflow-x: hidden;

Fixes #54.
Updates #24.